Relocate signal and projectrepo under internal/collector.
This allows the eventual simple reuse of signal collection across
binaries.
