+++ cmd/collect_signals/github/factory.go
@@ -20,7 +20,7 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/ossf/criticality_score/cmd/collect_signals/projectrepo"
+	"github.com/ossf/criticality_score/internal/collector/projectrepo"
 	"github.com/ossf/criticality_score/internal/githubapi"
 )
