+++ cmd/collect_signals/main.go
@@ -35,8 +35,8 @@ import (
 	"github.com/ossf/criticality_score/cmd/collect_signals/depsdev"
 	"github.com/ossf/criticality_score/cmd/collect_signals/github"
 	"github.com/ossf/criticality_score/cmd/collect_signals/githubmentions"
-	"github.com/ossf/criticality_score/cmd/collect_signals/projectrepo"
 	"github.com/ossf/criticality_score/cmd/collect_signals/result"
+	"github.com/ossf/criticality_score/internal/collector/projectrepo"
 	"github.com/ossf/criticality_score/internal/githubapi"
 	"github.com/ossf/criticality_score/internal/infile"
 	log "github.com/ossf/criticality_score/internal/log"
