+++ internal/collector/signal/signal.go
