+++ internal/collector/signal/source.go
@@ -17,7 +17,7 @@ package signal
 import (
 	"context"
 
-	"github.com/ossf/criticality_score/cmd/collect_signals/projectrepo"
+	"github.com/ossf/criticality_score/internal/collector/projectrepo"
 )
 
 // A Source is used to get a set of signals for a given project repository.
